edge-inference/requirements.txt
torch>=2.1.0,<2.2.0
torchvision>=0.16.0,<0.17.0
ultralytics>=8.0.0,<9.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
